+++ src/private/package/GetNovaPackageMetadataList.ps1
@@ -5,22 +5,61 @@ function Get-NovaPackageMetadataList {
         [Parameter(Mandatory)][pscustomobject]$ProjectInfo
     )
 
-    $packageSettings = $ProjectInfo.Package
-    $configuredPackageTypes = if ($packageSettings -is [System.Collections.IDictionary]) {
-        @($packageSettings['Types'])
+    $packageTypeList = @(Get-NovaConfiguredPackageTypeList -PackageSettings $ProjectInfo.Package)
+    $includeLatest = Test-NovaPackageLatestEnabled -PackageSettings $ProjectInfo.Package
+
+    return @(
+    foreach ($packageType in $packageTypeList) {
+        Get-NovaPackageMetadata -ProjectInfo $ProjectInfo -PackageType $packageType
+        if ($includeLatest) {
+            Get-NovaPackageMetadata -ProjectInfo $ProjectInfo -PackageType $packageType -Latest
+        }
+    }
+    )
+}
+
+function Get-NovaConfiguredPackageTypeList {
+    [CmdletBinding()]
+    param(
+        [AllowNull()]$PackageSettings
+    )
+
+    $configuredPackageTypes = if ($PackageSettings -is [System.Collections.IDictionary]) {
+        @($PackageSettings['Types'])
     }
     else {
-        @($packageSettings.Types)
+        @($PackageSettings.Types)
     }
 
     $packageTypeList = @($configuredPackageTypes | Where-Object {$_})
-    if ($packageTypeList.Count -eq 0) {
-        $packageTypeList = @('NuGet')
+    if ($packageTypeList.Count -gt 0) {
+        return $packageTypeList
     }
 
-    return @(
-    $packageTypeList | ForEach-Object {
-        Get-NovaPackageMetadata -ProjectInfo $ProjectInfo -PackageType $_
-    }
+    return @('NuGet')
+}
+
+function Test-NovaPackageLatestEnabled {
+    [CmdletBinding()]
+    param(
+        [AllowNull()]$PackageSettings
     )
+
+    if ($PackageSettings -is [System.Collections.IDictionary]) {
+        if ( $PackageSettings.Contains('Latest')) {
+            return [bool]$PackageSettings['Latest']
+        }
+
+        return $false
+    }
+
+    if ($null -eq $PackageSettings) {
+        return $false
+    }
+
+    if ($PackageSettings.PSObject.Properties.Name -contains 'Latest') {
+        return [bool]$PackageSettings.Latest
+    }
+
+    return $false
 }