+++ src/private/shared/GetNovaResolvedProjectPackageSettings.ps1
@@ -16,10 +16,12 @@ function Get-NovaResolvedProjectPackageSettings {
     Set-NovaPackageSettingDefault -PackageSettings $packageSettings -Name 'FileNamePattern' -Value "$( $packageSettings['Id'] )*" -TreatWhitespaceAsMissing
     Set-NovaPackageSettingDefault -PackageSettings $packageSettings -Name 'Authors' -Value $ManifestSettings['Author']
     Set-NovaPackageSettingDefault -PackageSettings $packageSettings -Name 'Description' -Value $ProjectData['Description'] -TreatWhitespaceAsMissing
+    Set-NovaPackageSettingDefault -PackageSettings $packageSettings -Name 'Latest' -Value $false
     Set-NovaPackageSettingDefault -PackageSettings $packageSettings -Name 'Repositories' -Value @()
     Set-NovaPackageSettingDefault -PackageSettings $packageSettings -Name 'Headers' -Value ([ordered]@{})
     Set-NovaPackageSettingDefault -PackageSettings $packageSettings -Name 'Auth' -Value ([ordered]@{})
 
+    $packageSettings['Latest'] = [bool]$packageSettings['Latest']
     $packageSettings['Repositories'] = @($packageSettings['Repositories'])
     $packageSettings['Headers'] = [ordered]@{} + $packageSettings['Headers']
     $packageSettings['Auth'] = [ordered]@{} + $packageSettings['Auth']